backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
langchain==0.1.9
langchain-openai==0.0.6
langchain-anthropic==0.0.2
langchain-google-genai==0.0.8
langgraph==0.0.26
langsmith==0.1.0
azure-devops==7.1.0b4
redis==5.0.1
asyncpg==0.29.0
neo4j==5.17.0
pydantic==2.5.3
python-dotenv==1.0.0
websockets==12.0
httpx==0.26.0
pyyaml==6.0.1
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
pytest-xdist==3.5.0
pytest-mock==3.12.0
aiofiles==23.2.1
